+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.2] - 2024-12-24
+
+### Fixed
+
+- Fixed Windows dual screen not working when both feeds use same input
+- Mirror left feed to right display when using same camera input (avoids Windows camera lock)
+
 ## [1.5.1] - 2024-12-24
 
 ### Fixed